+++ src/lib/queries/index.ts
@@ -89,6 +89,8 @@ export {
   useSeedDefaultQuestions,
   useGlobeMonitorSubmissionsByMonth,
   useGlobeMonitorSubmission,
+  useCreateGlobeMonitorSubmission,
+  useUserSubmissionsCountToday,
 } from "./globeMonitor";
 
 // Bookings